refactor: StickyMenu 애니메이션을 useScroll 훅으로 분리
